Move inner classes from IODevice to own classes files
