+++ src/EasyEPlanner.csproj
@@ -112,8 +112,11 @@
     <Compile Include="Configuration\ConfigurationChecker.cs" />
     <Compile Include="Configuration\DeviceSynchronizeService.cs" />
     <Compile Include="Device\IDevice.cs" />
+    <Compile Include="Device\IIOChannel.cs" />
     <Compile Include="Device\IIODevice.cs" />
+    <Compile Include="Device\IOChannel.cs" />
     <Compile Include="Device\IODevices\CAM.cs" />
+    <Compile Include="Device\IOLinkSize.cs" />
     <Compile Include="Device\Parameter.cs" />
     <Compile Include="Device\Property.cs" />
     <Compile Include="Device\RuntimeParameter.cs" />
